Move `datastructures` into `retrospective`
