+++ src/flowrep/api/schemas.py
@@ -14,21 +14,6 @@ from flowrep.base_models import PythonReference as PythonReference
 from flowrep.base_models import RecipeElementType as RecipeElementType
 from flowrep.base_models import RestrictedParamKind as RestrictedParamKind
 from flowrep.compiler.source import RenderedSource as RenderedSource
-from flowrep.datastructures import NOT_DATA as NOT_DATA
-from flowrep.datastructures import AtomicData as AtomicData
-from flowrep.datastructures import CompositeData as CompositeData
-from flowrep.datastructures import DagData as DagData
-from flowrep.datastructures import FlowControlData as FlowControlData
-from flowrep.datastructures import ForEachData as ForEachData
-from flowrep.datastructures import IfData as IfData
-from flowrep.datastructures import InputDataPort as InputDataPort
-from flowrep.datastructures import InputDataPorts as InputDataPorts
-from flowrep.datastructures import NodeData as NodeData
-from flowrep.datastructures import NotData as NotData
-from flowrep.datastructures import OutputDataPort as OutputDataPort
-from flowrep.datastructures import OutputDataPorts as OutputDataPorts
-from flowrep.datastructures import TryData as TryData
-from flowrep.datastructures import WhileData as WhileData
 from flowrep.edge_models import Edges as Edges
 from flowrep.edge_models import InputEdges as InputEdges
 from flowrep.edge_models import InputSource as InputSource
@@ -48,4 +33,19 @@ from flowrep.prospective.union_types import RecipeDiscrimination as RecipeDiscri
 from flowrep.prospective.union_types import Recipes as Recipes
 from flowrep.prospective.while_recipe import WhileRecipe as WhileRecipe
 from flowrep.prospective.workflow_recipe import WorkflowRecipe as WorkflowRecipe
+from flowrep.retrospective.datastructures import NOT_DATA as NOT_DATA
+from flowrep.retrospective.datastructures import AtomicData as AtomicData
+from flowrep.retrospective.datastructures import CompositeData as CompositeData
+from flowrep.retrospective.datastructures import DagData as DagData
+from flowrep.retrospective.datastructures import FlowControlData as FlowControlData
+from flowrep.retrospective.datastructures import ForEachData as ForEachData
+from flowrep.retrospective.datastructures import IfData as IfData
+from flowrep.retrospective.datastructures import InputDataPort as InputDataPort
+from flowrep.retrospective.datastructures import InputDataPorts as InputDataPorts
+from flowrep.retrospective.datastructures import NodeData as NodeData
+from flowrep.retrospective.datastructures import NotData as NotData
+from flowrep.retrospective.datastructures import OutputDataPort as OutputDataPort
+from flowrep.retrospective.datastructures import OutputDataPorts as OutputDataPorts
+from flowrep.retrospective.datastructures import TryData as TryData
+from flowrep.retrospective.datastructures import WhileData as WhileData
 from flowrep.subgraph_validation import ProspectiveOutputEdges as ProspectiveOutputEdges
