+++ src/flowrep/api/tools.py
@@ -8,10 +8,10 @@ for downstream library developers.
 from flowrep.compiler.source import flowrep2python as flowrep2python
 from flowrep.converters.python_workflow_definition import flowrep2pwd as flowrep2pwd
 from flowrep.converters.python_workflow_definition import pwd2flowrep as pwd2flowrep
-from flowrep.datastructures import recipe2data as recipe2data
 from flowrep.parsers.atomic_parser import atomic as atomic
 from flowrep.parsers.atomic_parser import parse_atomic as parse_atomic
 from flowrep.parsers.workflow_parser import parse_workflow as parse_workflow
 from flowrep.parsers.workflow_parser import workflow as workflow
+from flowrep.retrospective.datastructures import recipe2data as recipe2data
 from flowrep.storage import LexicalBagBrowser as LexicalBagBrowser
 from flowrep.wfms import run_recipe as run_recipe
