+++ src/flowrep/retrospective/datastructures.py
