+++ tests/unit/test_storage.py
@@ -9,7 +9,8 @@ import tempfile
 import unittest
 from unittest import mock
 
-from flowrep import datastructures, storage, storage_widget, wfms
+from flowrep import storage, storage_widget, wfms
+from flowrep.retrospective import datastructures
 
 from flowrep_static import library
